[refactor] id 에 int 대신 Long 자료형을 사용

+++ src/main/java/com/codessquad/qna/web/question/Question.java
@@ -1,12 +1,12 @@
 package com.codessquad.qna.web.question;
 
 public class Question {
-    int questionId;
+    long questionId;
     String writer;
     String title;
     String contents;
 
-    public int getQuestionId() {
+    public long getQuestionId() {
         return questionId;
     }
